code/requirements.txt
mysql-connector-python=3.0.0
bottle=0.12.15